+++ domutils/radar_tools/make_radar_fst.py
@@ -1,36 +1,71 @@
 import dask
 
+#identify logger names based on package hierarchy
+logging_basename = 'domutils.radar_tools'
+
 
 @dask.delayed
 def dask_to_fst(*args, **kwargs):
+    import sys
+    import logging
+    from dask.distributed import get_worker
+
+    #logger name is the same for all workers
+    logger = logging.getLogger(logging_basename)
+
+    #add handlers if none are present for this worker
+    if not len(logger.handlers):
+        command_line_args = args[2]
+        logger.setLevel(command_line_args.log_level)
+        logging.captureWarnings(True)
+        #handlers
+        worker_id = str(get_worker().id).lower()
+        stream_handler = logging.StreamHandler(sys.stdout)
+        file_handler = logging.FileHandler('logs/'+worker_id, 'w')
+        #levels
+        stream_handler.setLevel(command_line_args.log_level)
+        file_handler.setLevel(command_line_args.log_level)
+        #format
+        formatter_stream = logging.Formatter(worker_id+'    %(message)s')
+        stream_handler.setFormatter(formatter_stream)
+        formatter_file= logging.Formatter('%(asctime)s - %(name)s in %(funcName)s - %(levelname)s - %(message)s')
+        file_handler.setFormatter(formatter_file)
+        #add handlers
+        logger.addHandler(stream_handler)
+        logger.addHandler(file_handler)
+
     return to_fst(*args, **kwargs)
+
 
 def to_fst(valid_date, fst_template, args):
     #output data to std file
 
     import os
-    import warnings
+    import time
     import copy
+    import logging
     import numpy as np
     import rpnpy.librmn.all as rmn
     from rpnpy.rpndate import RPNDate
     from domutils import radar_tools
+    import domutils._py_tools as dpy
 
+    logger = logging.getLogger(logging_basename)
+    logger.info('to_fst starting to process date: '+str(valid_date))
 
     #output filename and directory
     output_file = args.output_dir + valid_date.strftime(args.fst_file_struc)
     #if in complete mode and file exists, return and test next one
     if (os.path.isfile(output_file) and args.complete_dataset):
-        print(output_file+ ' exists and complete_dataset=True. Skipping to the next.')
+        logger.info(output_file+ ' exists and complete_dataset=True. Skipping to the next.')
         return np.array([1])
     elif os.path.isfile(output_file):
         #file exists but we are not completing a dataset erase file before making a new one
         os.remove(output_file)
     else:
-        #we will create or overwrite the file make sure directory exists
+        #we will create or overwrite the file; before that make sure directory exists
         this_fst_dir = os.path.dirname(output_file)
-        if not os.path.isdir(this_fst_dir):
-            os.makedirs(this_fst_dir)
+        dpy.parallel_mkdir(this_fst_dir)
     
     #get destination grid from PR template
     dest_lon = fst_template['lon']
@@ -49,11 +84,10 @@ def to_fst(valid_date, fst_template, args):
                                                 dest_lon=dest_lon,
                                                 dest_lat=dest_lat,
                                                 median_filt=args.median_filt,
-                                                smooth_radius=args.smooth_radius,
-                                                verbose=1)
+                                                smooth_radius=args.smooth_radius)
 
         if desired_quantity == 'accumulation':
-            warnings.warn('!!!convert mm to m since PR quantity is outputted!!!')
+            logger.warning('!!!convert mm to m since PR quantity is outputted!!!')
             dat_dict['accumulation'] /= 1e3
         data_quantity_name  = desired_quantity
         data_date_name      = 'end_date'
@@ -77,14 +111,13 @@ def to_fst(valid_date, fst_template, args):
                                                  dest_lon=dest_lon,
                                                  dest_lat=dest_lat,
                                                  median_filt=args.median_filt,
-                                                 smooth_radius=args.smooth_radius,
-                                                 verbose=1)
+                                                 smooth_radius=args.smooth_radius)
         data_quantity_name  = desired_quantity
         data_date_name      = 'valid_date'
 
     #if we got nothing, fill output with nodata and zeros
     if dat_dict is None:
-        warnings.warn('no data found or file unreadeable, I observations are set to -9999. with quality index = 0.')
+        logger.warning('no data found or file unreadeable, I observations are set to -9999. with quality index = 0.')
         expected_shape  = dest_lat.shape
         precip_rate     = np.full(expected_shape, -9999.)
         quality_index   = np.zeros(expected_shape)
@@ -111,6 +144,7 @@ def to_fst(valid_date, fst_template, args):
     cmc_timestamp = date_obj.datev
 
     #open fst file
+    logger.info('writing ' + output_file)
     iunit = rmn.fstopenall(output_file,rmn.FST_RW)
 
     #write >> ^^ to output file
@@ -148,7 +182,7 @@ def to_fst(valid_date, fst_template, args):
     #close file
     rmn.fstcloseall(iunit)
 
-    print('Done writing ' + output_file)
+    logger.info('Done writing ' + output_file)
 
     #make a figure for this std file if the argument figure_dir was provided
     if args.figure_dir is not None:
@@ -158,6 +192,7 @@ def to_fst(valid_date, fst_template, args):
                                    fig_format=args.figure_format)
 
     return np.array([1])
+
 
 
 def parse_num(arg, dtype='int'):
@@ -200,14 +235,15 @@ def to_datetime(time_str):
 
 
 def make_fst(t0, tf, dt, args):
-    """
-    read odim H5, manipulate it, and write to fst
+    """ read odim H5, manipulate it, and write to fst
+
+    depending on the number of cpus, serial execution or parallel execution with dask will be chosen 
     """
 
     import os
     import datetime
     import glob
-
+    import logging
     import numpy as np
     import dask
     import dask.distributed
@@ -215,12 +251,16 @@ def make_fst(t0, tf, dt, args):
 
     from domcmc import fst_tools
 
+
+    #logging
+    logger = logging.getLogger(logging_basename)
+
     #make list of dates where radar data is needed
     t_len = (tf-t0) + datetime.timedelta(seconds=1)    #+ 1 second for inclusive end point
     elasped_seconds = t_len.days*3600.*24. + t_len.seconds
     date_list = [t0 + datetime.timedelta(seconds=x) for x in np.arange(0,elasped_seconds,dt)]
 
-    print('getting output domain from: ', args.sample_pr_file)
+    logger.info('getting output domain from: '+ args.sample_pr_file)
     fst_template = fst_tools.get_data(args.sample_pr_file, var_name='PR', latlon=True)
     if fst_template is None:
         raise ValueError('Problem getting PR from: ',args.sample_pr_file )
@@ -228,16 +268,17 @@ def make_fst(t0, tf, dt, args):
     #if only 1 cpu, dask is not used
     # makes for easier debugging 
     if args.ncores == 1 :
+        #serial execution
+        logger.info('Launching SERIAL execution of code')
         for this_date in date_list:
             to_fst(this_date, fst_template, args)
-            exit()
     else :
         #parallel conversion with dask
+        logger.info('Launching PARALLEL execution of code with dask')
 
         #open dask client for parallel execution
         #   level 40 hides dask warnings which are numerous and annoying 
         client = dask.distributed.Client(processes=True, threads_per_worker=1,n_workers=args.ncores,silence_logs=40)
-        #client = Client(..., silence_logs='error')
 
         # My functions returns 1 if success
         sample = np.array([1])
@@ -257,7 +298,7 @@ def make_fst(t0, tf, dt, args):
 
 
 
-if __name__ == "__main__":     
+def main():
     """ module for making a CMC "standard" file containing radar data
 
     It is intended to be used as a script, for example:
@@ -311,9 +352,17 @@ if __name__ == "__main__":
 
 
 
+    import os
+    from os import linesep as newline
+    import sys
     import argparse
+    import time
     import datetime
-    import sys,os
+    import logging
+    import domutils._py_tools as dpy
+
+    #keep track of runtime
+    time_start = time.time()
 
     #parse arguments
     desc="read radar H5 files, interpolate/smooth and write to FST"
@@ -334,9 +383,19 @@ if __name__ == "__main__":
     parser.add_argument("--complete_dataset" , type=str,   default='False',help="Skip existing files, default is to clobber them")
     parser.add_argument("--median_filt"      , type=str,   default='None', help="box size (pixels) for median filter")
     parser.add_argument("--smooth_radius"    , type=str,   default='None', help="radius (km) where radar data be smoothed")
-    parser.add_argument("--figure_dir"       , type=str,   default='None', help="If provided, a figure will be created for each std file created")
+    parser.add_argument("--figure_dir"       , type=str,   default='no_figures', help="If provided, a figure will be created for each std file created")
     parser.add_argument("--figure_format"    , type=str,   default='gif',  help="File format of figure ")
+    parser.add_argument("--log_level"        , type=str,   default='INFO', help="minimum level of messages printed to stdout and in log files ")
     args = parser.parse_args()
+
+
+    #add trailling / to all directories
+    args.radar_data_dir += '/'
+    args.output_dir += '/'
+    if args.figure_dir == 'no_figures' or args.figure_dir == 'None':
+        args.figure_dir = None
+    else:
+        args.figure_dir += '/'
 
     #parse accum_len
     if args.accum_len is not None :
@@ -379,7 +438,58 @@ if __name__ == "__main__":
 
     dt = parse_num(args.output_dt) * 60. #convert dt to seconds
 
+
+    # logging is configured to write everytiing to stdout in addition to a log file
+    # in a 'logs' directory
+
+    #make sure 'logs' directory exists and is empty
+    if os.path.isdir('logs'):
+        os.system('rm -f ./logs/main.log')
+        os.system('rm -f ./logs/worker*')
+    else:
+        #no need for parallel stuff here but the function already exists and will get the job done.
+        dpy.parallel_mkdir('logs')
+    logging.captureWarnings(True)
+    logger = logging.getLogger(logging_basename)
+    logger.setLevel(args.log_level)
+    #handlers
+    stream_handler = logging.StreamHandler(sys.stdout)
+    file_handler = logging.FileHandler('logs/main.log', 'w')
+    #levels
+    stream_handler.setLevel(args.log_level)
+    file_handler.setLevel(args.log_level)
+    #format
+    formatter = logging.Formatter('%(asctime)s - %(name)s in %(funcName)s - %(levelname)s - %(message)s')
+    
+    stream_handler.setFormatter(formatter)
+    file_handler.setFormatter(formatter)
+    #add handlers
+    logger.addHandler(stream_handler)
+    logger.addHandler(file_handler)
+
+    #log header
+    logger.info('')
+    logger.info('')
+    logger.info('executing python script:  domutils.radar_tools.make_radar_fst.py')
+    logger.info('All logs printed to stdout can also be found in ./logs/')
+    logger.info('')
+    logger.info('')
+
+    logger.info('After parsing, input arguments are:')
+    for arg in vars(args):
+       logger.info(arg +' = '+ str(getattr(args, arg)))
+    logger.info('')
+    logger.info('')
+
+
     #make std files
     make_fst(t0, tf, dt, args)
 
+    #we are done
+    time_stop = time.time()
+    logger.info('Python code completed, Runtime was : '+str(time_stop-time_start)+' seconds')
+
+
+if __name__ == "__main__":     
+    main()